requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==14.5.14
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.4
django-toolbelt==0.0.1
gunicorn==19.2.1
MarkupSafe==0.23
oauthlib==0.7.2
psycopg2==2.6
python-twitter==2.2
requests==2.5.1
requests-oauthlib==0.4.2
static3==0.5.1
djangorestframework==3.0.5
